requirements.txt
flask
matplotlib
numpy
opencv-python
pandas
Pillow
PyQt5
requests
scikit-image
scipy
pytest
